backend/requirements.txt
fastapi
uvicorn
twilio
psycopg2-binary
python-dotenv
apscheduler